+++ src/jk_svg/_AttrMixinXY.py
@@ -5,7 +5,7 @@ import typing
 import jk_typing
 import jk_hwriter
 
-from .AbstractSVGElement import AbstractSVGElement
+